Appendix
In the appendix A, we provide a detailed introduction to the impact of interpolation on rotation equivariance.
In the appendix B, we prove the rotation equivariance of proposed convolution-group framework.
In the appendix C and D, we show how to derive Eq. (21a) and Eq. (21b).
In the appendix E, we provide a concrete usage example to guide readers on how to correctly apply our theory.